Fixed bugs and pages are more responsive
