+++ src/app/using-zcash/non-custodial-exchanges/page.tsx
@@ -12,10 +12,12 @@ export const metadata: Metadata = genMetadata({
 const DEXListingPage = () => {
   return (
     <div className="container mx-auto px-4">
-      <div className="flex justify-between items-center my-12">
-        <h1 className="text-4xl font-semibold">Centralized Swap Platforms</h1>
+      <div className="flex justify-between items-center my-12 flex-col imd:flex-row">
+        <h1 className="flex justify-center items-center text-2xl imd:text-3xl font-bold mb-6 imd:mb-0 text-center">
+          Centralized Swap Platforms
+        </h1>
         <Link
-          href="https://zechub.wiki/using-zcash/custodial-exchanges"
+          href="/using-zcash/custodial-exchanges"
           className="inline-flex py-2 px-4 text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800 rounded-lg"
         >
           {/* <a className='bg-blue-500 text-white py-2 px-4 rounded hover:bg-blue-600'> */}
@@ -27,7 +29,7 @@ const DEXListingPage = () => {
         ZecHub does not endorse any particular exchange service, please do your
         own research.
       </p>
-      <div className="grid grid-cols-1 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+      <div className="grid grid-cols-1 sm:grid-cols-1 imd:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
         {dexListingConfig.map((itm, i) => (
           <Card
             thumbnailImage={itm.image}
